REDSHOP-3214 Do not delete Super Administrator - Joomla! User

+++ component/admin/models/user_detail.php
@@ -257,12 +257,18 @@ class RedshopModelUser_detail extends RedshopModel
 				$juserIds = $db->loadRowList();
 
 				foreach ($juserIds as $juserId) {
-					if (!JFactory::getUser($juserId[0])->delete())
-					{
-						$this->setError($db->getErrorMsg());
+                    $jUser = JFactory::getUser($juserId[0]);
 
-						return false;
-					}
+                    // Do not delete Super Administrator user
+                    if (!$jUser->authorise('core.admin'))
+                    {
+                        if (!JFactory::getUser($juserId[0])->delete())
+                        {
+                            $this->setError($db->getErrorMsg());
+
+                            return false;
+                        }
+                    }
 				}
 			}
